Add parent inode support to inodeDiscarders
